+++ CHANGELOG.md
@@ -5,9 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## 0.10.1 - 2024-05-29
 
 ### Fixed
+
 - minor typing and formatting issues
 - use `Deno.Command` in release scripts